requirements.txt
streamlit
yfinance
pandas
numpy
plotly
google-generativeai
python-dotenv